yet another fix to the json deserializer, a bug causing \0's to be appended to buffers when remainders are passed back to previous callstack function...
in a nutshell, when splitting the current read in buffer, and the read in count was less than the buffer size, and the cut point was also less than the buffer size, then \0's could potentially show up in the split buffer which in turn was used as part of the next iteration through, e.g. walkNames or walkValue.